Create new section
Second iteration, my main purpose here is to try to group rules
shared among many elements (like background color, etc) into a class
and apply that class to the html element.
I putted a div as a placeholder for an img, because the text inside
is quite easier to manipulate than the "alt" text of an img element.